Get TweetFetcher working
This demanded some pretty major surgery. Superficially, some methods are
moved from DataSet to TweetFetcher so tests and other things using that
interface need to be updated. But more critically, it now makes sense to
stage ALL the tweet data in Elasticsearch.

Originally, we kept Tweet objects on DataSets as they kept in and did
metadata extraction on those as part of the data set pipeline process.
However, we're now fetching tweets one user at a time -- and those
times may be pretty separated, since TweetFetcher will ultimately be a
delayed job subject to Twitter rate limits. This means that tweets
can't be accumulated on DataSet, but they ARE all available in
Elasticsearch.

Therefore it makes sense to reuse our existing extractor features which
rehydrate tweets from Elasticsearch, and update it to also be used in
the case where only one DataSet is fed into an extractor. This results
in a small change to the Extractor API (they are now initialized with
one or more DataSets instead of with Tweets).

On the bright side, when we improve extractor performance on
Elasticsearch, that work should automatically apply to all cases.

Next up, providing for delayed jobs and rate limiting!

+++ spec/support/elasticsearch.rb
@@ -38,6 +38,8 @@ RSpec.configure do |config|
   def wipe_elasticsearch_data
     client = Elasticsearch::Client.new
     DataSet.all.each do |ds|
+      next unless client.indices.exists? index: ds.index_name
+
       client.delete_by_query index: ds.index_name
       client.indices.delete index: ds.index_name
     end
